feat: export process output options in cli
